requirements.txt
Django==1.8.18
django-admin-bootstrapped==2.5.7
django-markdownx==2.0.22
Markdown==2.6.9
Pillow==6.0.0
psycopg2==2.7.3.2
Pygments==2.2.0
pytz==2017.2
uWSGI==2.0.15
